requirements.txt
google-adk==0.5.0
google-generativeai>=0.8.0
google-genai>=0.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
